pkg/lightning:move mock file for lightning into lightning package
