+++ pkg/lightning/mock/glue.go
