+++ pkg/lightning/mock/glue_checkpoint.go
