+++ pkg/lightning/mock/storage.go
